feat: update Loan and LoanType entities to use Short for periodMonths and id, and add CreateLoanRequest and CreateLoanResponse DTOs

+++ domain/model/src/main/java/co/com/pragma/model/loan/Loan.java
@@ -15,7 +15,7 @@ public class Loan {
 	private BigDecimal amount;
 	private String identificationNumber;
 	private String email;
-	private int periodMonths;
+	private Short periodMonths;
 	private LoanStatus status;
 	private LoanType loanType;
 }
